tools/requirements.txt
yt-dlp
openai-whisper
torch
ffmpeg-python
